Move Futures to top level of cpu module
