+++ crates/cpu/src/futures/get.rs
