+++ crates/cpu/src/futures/setters.rs
