+++ crates/cpu/src/opcodes/shift.rs
@@ -1,7 +1,5 @@
-use crate::registers::{
-    futures::{CbOperation as Operation, Set},
-    Bits8, Shift as S,
-};
+use crate::futures::{CbOperation as Operation, Set};
+use crate::registers::{Bits8, Shift as S};
 use crate::{Access, Cpu};
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 use shared::Error;
